Agones2 (#232)
* use agones2 branch


* specify kubernetes as a requirement


* install k8s


* don’t cache deps


* setup google k8s config


* remove double requirement of google auth


* move auth code in a place where code gets executed


* try another attempt at gke connection


* add debug message


* try another place


* update parameters of get cluster


* add debug list pods


* one more try with this approach


* add name keyword


* use kubeconfig yaml file


* remove migration step for now


* print contexts


* load kubeconfgi from dict


* read config file in properly


* move kubeconfig


* try again


* locate kubeconfig file


* setup dev and staging clusters


* Merge branch 'master' into agones2

* tested terraform on dev

* updated agones version and output.tf

* outputs.tf*

* fix deploy.sh

* fix kubeconfig template

* attempt fixing missing game_ide

* try to build aimmo from branch

* use node 12

* try alternative to update node

* back to nvm attempt

* more commands for nvm

* added node_modules to .gcloudignore

* missing comma in requirements

* build worker wheel

* rebuilt dev cluster and disabled clusters setup

* get latest published beta aimmo

* enable clusters setup

* downgraded agones version

* use dev clusters

* get aimmo from branch and some fixes

* some ingress patches

* ingress v1beta1

* fix cluster name in deploy

* aimmo from branch in requirements

* ingress and role in terraform and fleet in deploy

* delete only ingress in deploy

* delete game creator rs in deploy

* added more permissions to game-creator

* update fleet to 20 replicas

* reduce fleet to 15

* cleanup game servers and services

* recreated dev and cleanup aimmo branch deployment

* updated k8s and agones

* updated staging cluster

+++ clusters_setup/deploy.py
@@ -21,7 +21,8 @@ def create_ingress_yaml(module_name):
     """
     Uses a template ingress yaml file in the same directory to construct a python yaml
     object. Replaces the ingress.global-static-ip-name annotation with the
-    correct module name (depending on which cluster this script is ran in).
+    correct module name (depending on which cluster this script is ran in). Similarly for
+    cors-allow-origin, it replaces it with the correct url for selected module.
 
     Also replaces the TLS certificate host name and the host for the ingress paths.
 
@@ -30,15 +31,57 @@ def create_ingress_yaml(module_name):
     """
     path = os.path.join(CURR_DIR, "ingress.yaml")
 
-    host_name = module_name + "-aimmo.codeforlife.education"
+    host_name = f"{module_name}-aimmo.codeforlife.education"
+
+    cors_origin = f"https://{module_name}-dot-decent-digit-629.appspot.com"
+    if module_name == "default":
+        cors_origin = f"https://www.codeforlife.education"
 
     with open(path) as yaml_file:
         content = yaml.safe_load(yaml_file.read())
         content["metadata"]["annotations"][
             "kubernetes.io/ingress.global-static-ip-name"
-        ] = (module_name + "-aimmo-ingress")
+        ] = f"{module_name}-aimmo-ingress"
+        content["metadata"]["annotations"][
+            "nginx.ingress.kubernetes.io/cors-allow-origin"
+        ] = cors_origin
         content["spec"]["tls"][0]["hosts"][0] = host_name
         content["spec"]["rules"][0]["host"] = host_name
+
+    return content
+
+
+def create_fleet_yaml(module_name, aimmo_version):
+    """
+    Uses a template fleet yaml file in the same directory to construct a python yaml
+    object. Replaces image version with the one received as a parameter. Also replaces
+    the GAME_API_URL environment variable with the correct url for the deployment
+    environment we are in.
+
+    :param module_name: The name of the environment we're in (ie. staging, dev).
+    :param aimmo_version: The game version we want to deploy.
+    :return: python object containing yaml with modified values.
+    """
+    path = os.path.join(CURR_DIR, "fleet.yaml")
+
+    game_api_url = (
+        f"https://{module_name}-dot-decent-digit-629.appspot.com/kurono/api/games/"
+    )
+    container_image = f"ocadotechnology/aimmo-game:{aimmo_version}"
+
+    with open(path) as yaml_file:
+        content = yaml.safe_load(yaml_file.read())
+
+        env_variables = content["spec"]["template"]["spec"]["template"]["spec"][
+            "containers"
+        ][0]["env"]
+        game_api_url_index = env_variables.index(
+            {"name": "GAME_API_URL", "value": "REPLACE_ME"}
+        )
+        env_variables[game_api_url_index]["value"] = game_api_url
+        content["spec"]["template"]["spec"]["template"]["spec"]["containers"][0][
+            "image"
+        ] = container_image
 
     return content
 
@@ -50,7 +93,8 @@ def create_creator_yaml(module_name, aimmo_version):
     Replaces the GAME_API_URL environment variable with the correct url for the
     deployment environment we are in
 
-    :param module_name: The name of the environment we're in (ie. staging, dev)
+    :param module_name: The name of the environment we're in (ie. staging, dev).
+    :param aimmo_version: The game version we want to deploy.
     :return: python object containing yaml with modified values.
     """
 
@@ -87,36 +131,27 @@ def create_creator_yaml(module_name, aimmo_version):
     return content
 
 
-def restart_pods(game_creator_yaml, ingress_yaml):
+def restart_pods(game_creator_yaml, ingress_yaml, fleet_yaml):
     """
     Restarts the kubernetes replication controllers, pods, services and ingresses
     in the 'default' namespace
 
     :param game_creator_yaml: The dict to create the aimmo game creator rc
     :param ingress_yaml: The dict to create the ingress
+    :param fleet_yaml: The dict to create the fleet
     """
     for rs in apps_api_instance.list_namespaced_replica_set("default").items:
-        apps_api_instance.delete_namespaced_replica_set(
-            body=kubernetes.client.V1DeleteOptions(),
-            name=rs.metadata.name,
-            namespace="default",
-        )
-    for rc in api_instance.list_namespaced_replication_controller("default").items:
-        api_instance.delete_namespaced_replication_controller(
-            body=kubernetes.client.V1DeleteOptions(),
-            name=rc.metadata.name,
-            namespace="default",
-        )
-    for pod in api_instance.list_namespaced_pod("default").items:
-        api_instance.delete_namespaced_pod(
-            body=kubernetes.client.V1DeleteOptions(),
-            name=pod.metadata.name,
-            namespace="default",
-        )
+        if rs.metadata.name == game_creator_yaml["metadata"]["name"]:
+            apps_api_instance.delete_namespaced_replica_set(
+                body=kubernetes.client.V1DeleteOptions(),
+                name=rs.metadata.name,
+                namespace="default",
+            )
     for service in api_instance.list_namespaced_service("default").items:
-        api_instance.delete_namespaced_service(
-            name=service.metadata.name, namespace="default"
-        )
+        if service.metadata.name.startswith("game-"):
+            api_instance.delete_namespaced_service(
+                name=service.metadata.name, namespace="default"
+            )
     for ingress in networking_api_instance.list_namespaced_ingress("default").items:
         networking_api_instance.delete_namespaced_ingress(
             name=ingress.metadata.name,
@@ -124,10 +159,50 @@ def restart_pods(game_creator_yaml, ingress_yaml):
             body=kubernetes.client.V1DeleteOptions(),
         )
 
+    fleets_to_delete = custom_objects_api_instance.list_namespaced_custom_object(
+        group="agones.dev",
+        version="v1",
+        namespace="default",
+        plural="fleets",
+    )["items"]
+    for fleet in fleets_to_delete:
+        name = fleet["metadata"]["name"]
+        custom_objects_api_instance.delete_namespaced_custom_object(
+            group="agones.dev",
+            version="v1",
+            namespace="default",
+            plural="fleets",
+            name=name,
+        )
+
+    game_servers_to_delete = custom_objects_api_instance.list_namespaced_custom_object(
+        group="agones.dev",
+        version="v1",
+        namespace="default",
+        plural="gameservers",
+    )["items"]
+    for game_server in game_servers_to_delete:
+        name = game_server["metadata"]["name"]
+        custom_objects_api_instance.delete_namespaced_custom_object(
+            group="agones.dev",
+            version="v1",
+            namespace="default",
+            plural="gameservers",
+            name=name,
+        )
+
     networking_api_instance.create_namespaced_ingress("default", ingress_yaml)
 
     apps_api_instance.create_namespaced_replica_set(
         body=game_creator_yaml, namespace="default"
+    )
+
+    custom_objects_api_instance.create_namespaced_custom_object(
+        group="agones.dev",
+        version="v1",
+        namespace="default",
+        plural="fleets",
+        body=fleet_yaml,
     )
 
 
@@ -142,17 +217,20 @@ def main(module_name):
     global api_instance
     global apps_api_instance
     global networking_api_instance
+    global custom_objects_api_instance
     api_instance = kubernetes.client.CoreV1Api()
     apps_api_instance = kubernetes.client.AppsV1Api()
     networking_api_instance = kubernetes.client.NetworkingV1beta1Api()
+    custom_objects_api_instance = kubernetes.client.CustomObjectsApi()
     aimmo_version = get_aimmo_version()
 
     ingress = create_ingress_yaml(module_name=module_name)
     game_creator_rs = create_creator_yaml(
         module_name=module_name, aimmo_version=aimmo_version
     )
+    fleet = create_fleet_yaml(module_name=module_name, aimmo_version=aimmo_version)
 
-    restart_pods(game_creator_rs, ingress)
+    restart_pods(game_creator_rs, ingress, fleet)
 
 
 if __name__ == "__main__":
